CodeSolutionAgency/requirements.txt
selenium
webdriver-manager
selenium_stealth
openai==1.51.2
docstring_parser==0.16
pydantic==2.8.2
datamodel-code-generator==0.26.1
deepdiff==6.7.1
termcolor==2.4.0
python-dotenv==1.0.1
rich==13.7.1
jsonref==1.1.0
instructor
pytest
agency-swarm==0.2.6
